Set up page map and unmap capabilities
