+++ athena/microkernel/src/arch/x86_64/execution_context.rs
@@ -3,7 +3,8 @@
 use core::arch::asm;
 use core::mem::MaybeUninit;
 
-use super::paging::RawFrame;
+use super::paging::{PageTable, RawFrame};
+use crate::kptr::KPtr;
 
 /// A generic runnable context.
 ///
@@ -18,8 +19,8 @@ pub struct ExecutionContext {
 }
 
 impl ExecutionContext {
-    pub fn addrspace(&self) -> RawFrame {
-        self.address_space
+    pub fn addrspace(&self) -> KPtr<PageTable<4>> {
+        unsafe { PageTable::from_l4_frame(self.address_space.into_kernel_unchecked()) }
     }
 
     pub unsafe fn uninit() -> Self {
